2020 June 16, I resumed using smsn after a 2-year hiatus. My graph had evolved without being git-synced, and it was behind on Josh's updates, such that my "changes" actually undid some of his changes.
I pored through the diff and discovered that most of my changes to public/ and universal/ were bad; they were entirely due to my not having his changes. This commit (for which I am writing this message), I will write all the changes to those repops that are good, but none of the others. All my changes to personal/ and observatorio/ appear to be good; I am writing those changes unfiltered. (There were no chnages to private/.)